Updates in response to review
* Switched to a general "directories not allowed" policy on
  transfer/delete/etc.
* Changed `list_directory` to list all things in subdirectories as well

+++ openpathsampling/utils/storage_handlers.py
@@ -18,12 +18,26 @@ class StorageHandler(ABC):
     @abstractmethod
     def store(self, storage_label, source_path):
         """Store the data in ``source_path`` at key ``storage_label``
+
+        Parameters
+        ----------
+        storage_label : str
+            The key to store the data at.
+        source_path : os.PathLike
+            The path to the data to store.
         """
         raise NotImplementedError()
 
     @abstractmethod
     def load(self, storage_label, target_path):
         """Load the data from ``storage_label`` into file at ``target_path``
+
+        Parameters
+        ----------
+        storage_label : str
+            The key to load the data from.
+        target_path : os.PathLike
+            The path to store the data at.
         """
         raise NotImplementedError()
 
@@ -44,16 +58,26 @@ class StorageHandler(ABC):
         os.remove, so this method can be overridden. (Example: moving ona
         file system is faster than copying.)
         """
+        if pathlib.Path(source_path).is_dir():
+            raise ValueError(f"'{source_path}' is a directory, and can't "
+                             "be transferred.")
         self.store(storage_label, source_path)
         os.remove(source_path)
 
     @abstractmethod
     def list_directory(self, storage_label):
+        """List all objects in subdirectories of the given storage label.
+        """
         raise NotImplementedError()
 
 
 class LocalFileStorageHandler(StorageHandler):
     """Concrete implementation of StorageHandler for local files.
+
+    Parameters
+    ----------
+    root : os.PathLike
+        The root directory for the storage handler.
     """
     def __init__(self, root):
         self.root = pathlib.Path(root)
@@ -72,21 +96,40 @@ class LocalFileStorageHandler(StorageHandler):
         shutil.copyfile(local_path, target_path)
 
     def delete(self, storage_label):
-        os.remove(self.root / storage_label)
+        obj = self.root / storage_label
+        if obj.is_dir():
+            raise ValueError(f"'{obj}' is a directory, and can't be "
+                             "deleted.")
+        else:
+            _logger.debug("Deleting file {str(obj)}")
+            os.remove(self.root / storage_label)
 
     def __contains__(self, storage_label):
-        return (self.root / storage_label).exists()
+        expected = self.root / storage_label
+        return expected.exists() and not expected.is_dir()
 
     def transfer(self, storage_label, source_path):
+        if pathlib.Path(source_path).is_dir():
+            raise ValueError(f"'{source_path}' is a directory, and can't "
+                             "be transferred.")
         shutil.move(source_path, self.root / storage_label)
 
     def list_directory(self, storage_label):
-        return [str(p.relative_to(self.root))
-                for p in (self.root / storage_label).iterdir()]
+        path = self.root / storage_label
+        if not path.is_dir():
+            raise ValueError(f"'{path}' is not a directory.")
+        return [
+            str((pathlib.Path(p[0]) / subp).relative_to(self.root))
+            for p in os.walk(path)
+            for subp in p[2]
+        ]
 
 
 class MemoryStorageHandler(StorageHandler):
-    """Useful in testing."""
+    """In-memory storage handler.
+
+    Useful in testing.
+    """
     def __init__(self):
         self._data = {}
 
@@ -109,6 +152,8 @@ class MemoryStorageHandler(StorageHandler):
         # special case because the empty path becomes '.' as a string
         if storage_label == pathlib.Path(""):
             storage_label = ""
+        elif not storage_label.endswith("/"):
+            storage_label += "/"
 
         return [key for key in self._data
                 if str(key).startswith(str(storage_label))]
